requirements.txt
google-generativeai
beautifulsoup4
requests
pandas
matplotlib
python-dotenv
rich
pytest
feedparser
